(wip) fix ourorobors_opengv packaging
