+++ src/ouroboros/__init__.py
@@ -13,6 +13,8 @@ from ouroboros.vlc_db import (
     VlcImage,
     VlcImageMetadata,
     VlcPose,
+    invert_pose,
+    pose_from_quat_trans,
 )
 from ouroboros.vlc_server.vlc_server import VlcServer, VlcServerConfig
